+++ Js2IL.Tests/BinaryOperator/GeneratorTests.cs
@@ -37,6 +37,9 @@ namespace Js2IL.Tests.BinaryOperator
         public Task BinaryOperator_EqualBoolean() { var testName = nameof(BinaryOperator_EqualBoolean); return GenerateTest(testName); }
 
         [Fact]
+        public Task BinaryOperator_EqualMethodReturn() { var testName = nameof(BinaryOperator_EqualMethodReturn); return GenerateTest(testName); }
+
+        [Fact]
         public Task BinaryOperator_ExpNumberNumber() { var testName = nameof(BinaryOperator_ExpNumberNumber); return GenerateTest(testName); }
 
         [Fact]